% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipl5-package.R
\docType{package}
\name{bipl5}
\alias{bipl5}
\title{bipl5: Constructing Reactive Calibrated Axes Biplots}
\description{
A modern view on PCA biplot with calibrated axes. Create PCA biplots
rendered in HTML with significant reactivity embedded on the plot.
Furthermore, the traditional biplot view is enhanced by translated axes
with interclass kernel densities superimposed. For more information on
biplots, see Gower, J.C., Lubbe, S. and le Roux, N.J.
(2011, ISBN: 978-0-470-01255-0)
}
\section{Details}{


\tabular{ll}{
Package: \tab bipl5 \cr
Type: \tab Package \cr
Version: \tab 1.1 \cr
Date: \tab 18-12-2023\cr
License: \tab MIT \cr
}
}

\section{Author(s)}{


\itemize{
\item{Ruan Buys (Maintainer)}

\item{Carel van der Merwe}

\item{Delia Sandilands (contributer)}

\item{Sugnet Lubbe (contributer)}

}
}

\section{Core Functions}{

\itemize{
\item{\link{wrap_bipl5}}

\item{\link{extract}}

\item{\link{append_payload}}

\item{\link{remove_payload}}

\item{\link{plot.bipl5_biplot}}
}
}

\section{Code Availability}{


The newest version of the package can be obtained on
GitHub: \url{https://github.com/RuanBuys/bipl5}
}

