% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translate_Helpers.R
\name{colorpal}
\alias{colorpal}
\title{bpl5 default color scales}
\usage{
colorpal(number = 16)
}
\arguments{
\item{number}{Number of distinct colors to return}
}
\value{
A vector of colors
}
\description{
bpl5 default color scales
}
\examples{
colorpal()
}
