% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmeasures.R
\name{FMbiplot}
\alias{FMbiplot}
\title{Determine various measures of fit for the PCA biplot}
\usage{
FMbiplot(x)
}
\arguments{
\item{x}{An object of class \code{bipl5}}
}
\value{
A list returned invisibly containing the following fit measures:
\itemize{
\item Cumulative Predictivity
\item Marginal Predictivity for the first three principal components
\item Marginal Adequacy for the first three principal components
\item Overall quality of display
}
}
\description{
Print various measures of fit of the biplot display to the console
}
\examples{
x<-PCAbiplot(iris[,-5])
FMbiplot(x)
}
