% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_secondary_biplots.R
\name{insert_Z_coo_payload}
\alias{insert_Z_coo_payload}
\title{Add the main observations to the payload}
\usage{
insert_Z_coo_payload(payload, x, p_ly_pch, Col, visible = TRUE)
}
\arguments{
\item{payload}{List containing data and layout attributes for a plotly graph}

\item{x}{Object from biplotEZ package}

\item{p_ly_pch}{Plotting characters of the observations}

\item{Col}{Colors of the observations to be plotted}

\item{visible}{Whether points should be visible on graph. True by default}
}
\description{
Add the main observations to the payload
}
