% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplotEZ_plotter.R
\name{plot_bipl5.PCO}
\alias{plot_bipl5.PCO}
\title{Plot a PCO biplot from the biplotEZ package}
\usage{
\method{plot_bipl5}{PCO}(x)
}
\arguments{
\item{x}{An object of class biplotEZ::biplot}
}
\value{
A plotly graph
}
\description{
Plot a PCO biplot from the biplotEZ package
}
\examples{
x<-biplot(iris[,1:4]) |> PCO(dist.func = sqrtManhattan) |> plot_bipl5()
}
