% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_class_bipl5.R
\name{predict.bipl5}
\alias{predict.bipl5}
\title{Method to obtain the predicted or fitted values of the biplot}
\usage{
\method{predict}{bipl5}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bipl5} from which predicted values are to be obtained}

\item{...}{Additional arguments to be passes to the kable function, contained in a list named kable.args}
}
\value{
The function invisibly returns the predicted values of the biplot display,
and outputs the predicted values via the \code{\link[knitr]{kable}} function
}
\description{
Extract the fitted values of the biplot display
}
\examples{
kable.args<-list()
kable.args$format<-"pipe"
x<-PCAbiplot(iris[,-5])
predict(x,kable.args)
}
