% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_class_bipl5.R
\name{print.bipl5}
\alias{print.bipl5}
\title{Default print method for an object of class \code{bipl5}}
\usage{
\method{print}{bipl5}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bipl5}}

\item{...}{Additional parameters}
}
\value{
The object is returned invisibly
}
\description{
Default print method for an object of class \code{bipl5}
}
\examples{
x<-PCAbiplot(iris[,1:4],group=iris[,5])
print.bipl5(x)
}
