% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bipl5.R
\name{print.bipl5_biplot}
\alias{print.bipl5_biplot}
\title{Print a bipl5_biplot object as a tree diagram}
\usage{
\method{print}{bipl5_biplot}(x, ...)
}
\arguments{
\item{x}{A \code{bipl5_biplot} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print a bipl5_biplot object as a tree diagram
}
