% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bipl5.R
\name{print.bipl5_data}
\alias{print.bipl5_data}
\title{Print a bipl5_data object}
\usage{
\method{print}{bipl5_data}(x, ...)
}
\arguments{
\item{x}{A \code{bipl5_data} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print a bipl5_data object
}
