% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bipl5.R
\name{print.bipl5_fitmeasures}
\alias{print.bipl5_fitmeasures}
\title{Print a bipl5_fitmeasures object}
\usage{
\method{print}{bipl5_fitmeasures}(x, ...)
}
\arguments{
\item{x}{A \code{bipl5_fitmeasures} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print a bipl5_fitmeasures object
}
