% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bipl5.R
\name{print.bipl5_mdsDisplay}
\alias{print.bipl5_mdsDisplay}
\title{Print a bipl5_mdsDisplay object}
\usage{
\method{print}{bipl5_mdsDisplay}(x, ...)
}
\arguments{
\item{x}{A \code{bipl5_mdsDisplay} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print a bipl5_mdsDisplay object
}
