% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bipl5.R
\name{print.bipl5_payload}
\alias{print.bipl5_payload}
\title{Print a bipl5_payload object}
\usage{
\method{print}{bipl5_payload}(x, ...)
}
\arguments{
\item{x}{A \code{bipl5_payload} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print a bipl5_payload object
}
