% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bipl5.R
\name{remove_mdsDisplay}
\alias{remove_mdsDisplay}
\alias{remove_mdsDisplay.bipl5_biplot}
\title{Remove a mdsDisplay from a bipl5_biplot object}
\usage{
remove_mdsDisplay(object, mdsDisplay)

\method{remove_mdsDisplay}{bipl5_biplot}(object, mdsDisplay)
}
\arguments{
\item{object}{A \code{bipl5_biplot} object}

\item{mdsDisplay}{Unquoted name of the mdsDisplay to remove
(e.g. \code{mdsDisplay_13})}
}
\value{
A new \code{bipl5_biplot} without the removed mdsDisplay
}
\description{
Returns a new \code{bipl5_biplot} with the specified mdsDisplay (and its
corresponding fit table) removed.  At least one mdsDisplay must remain.
}
