% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods_class_bipl5.R
\name{summary.bipl5}
\alias{summary.bipl5}
\title{Obtain summary statistics of an \code{bipl5} object}
\usage{
\method{summary}{bipl5}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bipl5}}

\item{...}{Not applicable}
}
\value{
The object is returned invisibly
}
\description{
Obtain summary statistics of an \code{bipl5} object
}
\examples{
x<-PCAbiplot(iris[,-5])
summary(x)
}
