% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bipl5.R
\name{wrap_bipl5}
\alias{wrap_bipl5}
\title{Convert a biplotEZ object to a bipl5_biplot}
\usage{
wrap_bipl5(x)
}
\arguments{
\item{x}{A biplotEZ biplot object}
}
\value{
An object of class \code{bipl5_biplot}
}
\description{
Convert a biplotEZ object to a bipl5_biplot
}
