% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Translate_Helpers.R
\name{colorpal}
\alias{colorpal}
\title{bipl5 default color scales}
\usage{
colorpal(number = 16)
}
\arguments{
\item{number}{Integer - number of distinct colors to return. Ranges from
1 to 16.}
}
\value{
Character vector of default colors in \code{bipl5}. There are
sixteen unique colors defined.
}
\description{
bipl5 default color scales
}
\examples{
colorpal(number=7)
}
