% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting Symbols.R
\name{pch_to_plotly}
\alias{pch_to_plotly}
\title{Convert base plotting characters to plotly-compatible ones}
\usage{
pch_to_plotly(pch)
}
\arguments{
\item{pch}{Numeric vector of plotting characters}
}
\value{
Character vector of plotly pch's
}
\description{
Convert base plotting characters to plotly-compatible ones
}
