% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplotEZ plotter.R
\name{plot_bipl5.PCA}
\alias{plot_bipl5.PCA}
\title{Plot a PCA biplot from the biplotEZ package}
\usage{
\method{plot_bipl5}{PCA}(x)
}
\arguments{
\item{x}{An object of class biplotEZ::biplot}
}
\value{
A plotly graph
}
\description{
Plot a PCA biplot from the biplotEZ package
}
\examples{
x<-biplot(data = iris) |> PCA() |> plot_bipl5()
}
