% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplotEZ plotter.R
\name{plot_bipl5}
\alias{plot_bipl5}
\title{Plot object from the biplotEZ package in plotly}
\usage{
plot_bipl5(x)
}
\arguments{
\item{x}{An object of class 'biplot' from the biplotEZ package}
}
\value{
A plotly graph containing the biplot
}
\description{
Plot object from the biplotEZ package in plotly
}
\examples{
x<-biplot(data = iris) |> PCA() |> plot_bipl5()
x<-biplot(iris[,1:4]) |> CVA(classes=iris[,5]) |> plot_bipl5()
x<-biplot(iris[,1:4]) |> PCO(dist.func = sqrtManhattan) |> plot_bipl5()
}
