% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bipl5.R
\name{wrap_bipl5.CVA}
\alias{wrap_bipl5.CVA}
\title{Construct a bipl5_biplot from a CVA biplot}
\usage{
\method{wrap_bipl5}{CVA}(x)
}
\arguments{
\item{x}{An object of class \code{biplot} from the biplotEZ package with
CVA method applied.}
}
\value{
An object of class \code{c("bipl5_biplot", "cva")}
}
\description{
Builds mdsDisplays for the user's CV pair and available supplementary pairs,
along with a dropdown menu.  Fit measures are not yet computed for CVA
biplots and will be \code{NULL}.
Plotting is deferred to \code{\link{plot.bipl5_biplot}}.
}
\examples{
\dontrun{
library(biplotEZ)
bp <- biplot(iris[, 1:4]) |> CVA(classes = iris[, 5]) |> wrap_bipl5()
bp
plot(bp)
}
}
