% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_bipl5.R
\name{wrap_bipl5.PCA}
\alias{wrap_bipl5.PCA}
\title{Construct a bipl5_biplot from a PCA biplot}
\usage{
\method{wrap_bipl5}{PCA}(x)
}
\arguments{
\item{x}{An object of class \code{biplot} from the biplotEZ package with
PCA method applied.}
}
\value{
An object of class \code{bipl5_biplot}
}
\description{
Builds three payloads for the user's PC pair and two supplementary pairs,
along with fit measures.  The user's original PC pair is always first.
Plotting is deferred to \code{\link{plot.bipl5_biplot}}.
}
\examples{
\dontrun{
library(biplotEZ)
bp <- biplot(data = iris) |> PCA() |> wrap_bipl5()
bp
plot(bp)
}
}
